feat: Enhance relay management with group relay support, refactor UI components for improved drag-and-drop functionality
