+++ lib/src/features/modul/presentation/screen/modul_detail_screen.dart
@@ -5,6 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:lucide_icons_flutter/lucide_icons.dart';
 import 'package:pak_tani/src/core/config/app_config.dart';
+import 'package:pak_tani/src/core/theme/app_shadows.dart';
 import 'package:pak_tani/src/core/theme/app_theme.dart';
 import 'package:pak_tani/src/core/utils/modul_feature_helper.dart';
 import 'package:pak_tani/src/core/widgets/battery_status.dart';
@@ -286,7 +287,7 @@ class ModulDetailScreen extends StatelessWidget {
                     child: controller.isTitleExpanded.value
                         ? BlurryContainer(
                             key: const ValueKey('expandedDescription'),
-                            blur: 10,
+                            blur: 6,
                             padding: EdgeInsetsGeometry.all(0),
                             child: Container(
                               decoration: BoxDecoration(
@@ -304,49 +305,19 @@ class ModulDetailScreen extends StatelessWidget {
                               child: Column(
                                 crossAxisAlignment: CrossAxisAlignment.start,
                                 children: [
-                                  Stack(
-                                    children: [
-                                      // Outline hitam
-                                      Text(
-                                        controller.modul.value!.name,
-                                        style: AppTheme.h4.copyWith(
-                                          foreground: Paint()
-                                            ..style = PaintingStyle.stroke
-                                            ..strokeWidth = 1
-                                            ..color = Colors.black,
-                                        ),
-                                      ),
-                                      // Text putih di atas
-                                      Text(
-                                        controller.modul.value!.name,
-                                        style: AppTheme.h4.copyWith(
-                                          color: Colors.white,
-                                        ),
-                                      ),
-                                    ],
+                                  Text(
+                                    controller.modul.value!.name,
+                                    style: AppTheme.h4.copyWith(
+                                      color: Colors.white,
+                                      shadows: AppShadows.blackFade,
+                                    ),
                                   ),
-                                  Stack(
-                                    children: [
-                                      // Outline hitam untuk deskripsi
-                                      Text(
-                                        controller.modul.value!.descriptions ??
-                                            "",
-                                        style: AppTheme.text.copyWith(
-                                          foreground: Paint()
-                                            ..style = PaintingStyle.stroke
-                                            ..strokeWidth = 0.8
-                                            ..color = Colors.black,
-                                        ),
-                                      ),
-                                      // Text putih di atas
-                                      Text(
-                                        controller.modul.value!.descriptions ??
-                                            "",
-                                        style: AppTheme.text.copyWith(
-                                          color: Colors.white,
-                                        ),
-                                      ),
-                                    ],
+                                  Text(
+                                    controller.modul.value!.descriptions ?? "",
+                                    style: AppTheme.text.copyWith(
+                                      color: Colors.white,
+                                      shadows: AppShadows.blackFade,
+                                    ),
                                   ),
                                 ],
                               ),
